+++ R/PsFunctions.R
@@ -67,6 +67,7 @@ createPs <- function(cohortMethodData,
                      control = createControl(noiseLevel = "silent",
                                              cvType = "auto",
                                              seed = 1,
+                                             resetCoefficients = TRUE,
                                              tolerance = 2e-07,
                                              cvRepetitions = 10,
                                              startingVariance = 0.01)) {